+++ R/deployApp.R
@@ -23,6 +23,7 @@
 #' @param appFiles A character vector given relative paths to the files and
 #'   directories to bundle and deploy. The default, `NULL`, will include all
 #'   files in `appDir`, apart from any listed in an `.rscignore` file.
+#'   See [listBundleFiles()] for more details.
 #' @param appFileManifest An alternate way to specify the files to be deployed.
 #'   Should be a path to a file that contains the names of the files and
 #'   directories to deploy, one per line, relative to `appDir`.
@@ -244,20 +245,11 @@ deployApp <- function(appDir = getwd(),
     preDeploy(appPath)
   }
 
-  # figure out what kind of thing we're deploying
-  if (!is.null(contentCategory)) {
-    assetTypeName <- contentCategory
-  } else if (!is.null(appPrimaryDoc)) {
-    assetTypeName <- "document"
-  } else {
-    assetTypeName <- "application"
-  }
-
   appFiles <- standardizeAppFiles(appDir, appFiles, appFileManifest)
 
   if (isTRUE(lint)) {
     lintResults <- lint(appDir, appFiles, appPrimaryDoc)
-    showLintResults(appDir, lintResults, assetTypeName)
+    showLintResults(appDir, lintResults)
   }
 
   # determine the deployment target and target account info
@@ -277,7 +269,7 @@ deployApp <- function(appDir = getwd(),
   }
 
   # get the application to deploy (creates a new app on demand)
-  withStatus(paste0("Preparing to deploy ", assetTypeName), {
+  withStatus("Preparing to deploy", {
     application <- applicationForTarget(client, accountDetails, target, forceUpdate)
   })
 
@@ -301,13 +293,12 @@ deployApp <- function(appDir = getwd(),
   if (upload) {
     # create, and upload the bundle
     logger("Bundle upload started")
-    withStatus(paste0("Uploading bundle for ", assetTypeName, ": ",
-                     application$id), {
+    withStatus(paste0("Uploading bundle (", application$id, ")"), {
 
       # python is enabled on Connect but not on Shinyapps
       python <- getPythonForTarget(python, accountDetails)
       bundlePath <- bundleApp(target$appName, appDir, appFiles,
-                              appPrimaryDoc, assetTypeName, contentCategory, verbose, python,
+                              appPrimaryDoc, contentCategory, verbose, python,
                               condaMode, forceGeneratePythonEnvironment, quarto,
                               isCloudServer(accountDetails$server), metadata, image)
 
@@ -367,7 +358,7 @@ deployApp <- function(appDir = getwd(),
 
   if (length(bundle$id) > 0 && nzchar(bundle$id)) {
     displayStatus(paste0("Deploying bundle: ", bundle$id,
-                         " for ", assetTypeName, ": ", application$id,
+                         " (", application$id, ")",
                          " ...\n", sep = ""))
   }
 
@@ -383,12 +374,10 @@ deployApp <- function(appDir = getwd(),
   Sys.sleep(0.10)
 
   deploymentSucceeded <- if (is.null(response$code) || response$code == 0) {
-    displayStatus(paste0(capitalize(assetTypeName), " successfully deployed ",
-                         "to ", application$url, "\n"))
+    displayStatus(paste0("Successfully deployed to ", application$url, "\n"))
     TRUE
   } else {
-    displayStatus(paste0(capitalize(assetTypeName), " deployment failed ",
-                         "with error: ", response$error, "\n"))
+    displayStatus(paste0("Deployment failed with error: ", response$error, "\n"))
     FALSE
   }
 
@@ -416,7 +405,7 @@ deployApp <- function(appDir = getwd(),
 # deployApp() instead of being exposed to the user. Returns the path to the
 # bundle directory, whereas writeManifest() returns nothing and deletes the
 # bundle directory after writing the manifest.
-bundleApp <- function(appName, appDir, appFiles, appPrimaryDoc, assetTypeName,
+bundleApp <- function(appName, appDir, appFiles, appPrimaryDoc,
                       contentCategory, verbose = FALSE, python = NULL,
                       condaMode = FALSE, forceGenerate = FALSE, quarto = NULL,
                       isCloudServer = FALSE, metadata = list(), image = NULL) {
@@ -455,7 +444,6 @@ bundleApp <- function(appName, appDir, appFiles, appPrimaryDoc, assetTypeName,
       contentCategory = contentCategory,
       hasParameters = appMetadata$hasParameters,
       appPrimaryDoc = appMetadata$appPrimaryDoc,
-      assetTypeName = assetTypeName,
       users = users,
       condaMode = condaMode,
       forceGenerate = forceGenerate,